Several changes
Removed not needed requires in Windurst Wall Greeter Moogle.

Move login_campaign.lua to events folder, removed the table,
renamed function names and charVar names to be friendlier,
removed debug messages, changed messageSpecial magic number
to instead use textIDs.

login_campaign_data.lua now holds the login campaign tables.

moved onGameIn to player.lua, still needs to be fixed as
doesnt show messages, everything else works.

removed function from Zone.lua as is no longer being
triggered in there.
