+++ scripts/zones/Windurst_Walls/Zone.lua
@@ -10,7 +10,6 @@ require("scripts/globals/keyitems")
 require("scripts/globals/missions")
 require("scripts/globals/quests")
 require("scripts/globals/zone")
-require("scripts/globals/logincampaign")
 -----------------------------------
 
 function onInitialize(zone)
@@ -40,11 +39,6 @@ function onZoneIn(player, prevZone)
     end
 
     return cs
-end
-
-function afterZoneIn(player)
-    -- LOGIN CAMPAIGN
-    tpz.logincampaign.points(player)
 end
 
 function onConquestUpdate(zone, updatetype)
