+++ scripts/zones/Windurst_Walls/npcs/Greeter_Moogle.lua
@@ -3,26 +3,18 @@
 --  NPC: Greeter Moogle
 -- !pos NEED POS
 -----------------------------------
-local ID = require("scripts/zones/Windurst_Walls/IDs")
-require("scripts/globals/keyitems")
-require("scripts/globals/settings")
-require("scripts/globals/wsquest")
-require("scripts/globals/quests")
-require("scripts/globals/titles")
-require("scripts/globals/utils")
-require("scripts/globals/logincampaign")
-require("scripts/globals/npc_util")
+require("scripts/globals/events/login_campaign")
 -----------------------------------
 
 function onTrade(player, npc, trade)
 end
 
 function onTrigger(player, npc)
-    tpz.logincampaign.check(player)
+    tpz.events.loginCampaign.onTrigger(player)
 end
 
 function onEventUpdate(player, csid, option)
-    tpz.logincampaign.eventupdate(player, csid, option)
+    tpz.events.loginCampaign.onEventUpdate(player, csid, option)
 end
 
 function onEventFinish(player, csid, option)
